requirements.txt
python-telegram-bot==13.3
Scrapy==2.4.1
scrapy-rotating-proxies==0.6.2
requests==2.25.1
Scrapy-UserAgents==0.0.1
nltk==3.4.5
fasttext==0.9.2
sklearn==0.0
ftfy==5.9
numpy==1.20.1
python-dotenv==0.15.0